+++ metadata.rb
@@ -2,6 +2,6 @@ name 's3backup'
 maintainer 'Alexander Pyatkin'
 maintainer_email 'aspyatkin@gmail.com'
 description 'Setup a scheduled backup on Amazon S3'
-version '3.0.0'
+version '3.1.0'
 
-depends 'instance', '~> 2.0.1'
+gem 'mixlib-versioning'