+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2020-03-05
+
+### Added
+- `python` property is added to `s3backup_item` resource.
+
+### Changed
+- install `awscli` in a virtualenv.
+- no longer dependent on `instance` cookbook
+
 ## [3.0.0] - 2020-03-04
 
 ### Changed